% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyExample-package.R
\docType{package}
\name{MyExample}
\alias{MyExample}
\alias{MyExample-package}
\title{MyExample: This is my first R package}
\description{
This package was created as a way to learn how to write R packages.
}
