% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square}
\usage{
square(x)
}
\arguments{
\item{x}{Vector or numerical number.}
}
\value{
Square of x.
}
\description{
\code{square} returns the square term for a numerical number or vector.
}
\details{
More details
}
\examples{

square(3)

}
